+++ v2/models/emailfilter_options.go
@@ -3,5 +3,5 @@ package models
 const EmailfilterOptionsPath = "emailfilter/options/"
 
 type EmailfilterOptions struct {
-	DnsTimeout *int64 `json:"dns-timeout,omitempty"`
+	DnsTimeout *float64 `json:"dns-timeout,omitempty"`
 }
