+++ v2/models/firewall.schedule_recurring.go
@@ -3,10 +3,10 @@ package models
 const FirewallscheduleRecurringPath = "firewall.schedule/recurring/"
 
 type FirewallscheduleRecurring struct {
-	Color        *int64  `json:"color,omitempty"`
-	Day          *string `json:"day,omitempty"`
-	End          *string `json:"end,omitempty"`
-	FabricObject *string `json:"fabric-object,omitempty"`
-	Name         *string `json:"name,omitempty"`
-	Start        *string `json:"start,omitempty"`
+	Color        *float64 `json:"color,omitempty"`
+	Day          *string  `json:"day,omitempty"`
+	End          *string  `json:"end,omitempty"`
+	FabricObject *string  `json:"fabric-object,omitempty"`
+	Name         *string  `json:"name,omitempty"`
+	Start        *string  `json:"start,omitempty"`
 }
