+++ v2/models/firewall.wildcard_fqdn_group.go
@@ -3,7 +3,7 @@ package models
 const FirewallwildcardFqdnGroupPath = "firewall.wildcard-fqdn/group/"
 
 type FirewallwildcardFqdnGroup struct {
-	Color      *int64                             `json:"color,omitempty"`
+	Color      *float64                           `json:"color,omitempty"`
 	Comment    *string                            `json:"comment,omitempty"`
 	Member     *[]FirewallwildcardFqdnGroupMember `json:"member,omitempty"`
 	Name       *string                            `json:"name,omitempty"`
