+++ v2/models/firewall_DoS_policy6.go
@@ -8,22 +8,22 @@ type FirewallDoSPolicy6 struct {
 	Dstaddr   *[]FirewallDoSPolicy6Dstaddr `json:"dstaddr,omitempty"`
 	Interface *string                      `json:"interface,omitempty"`
 	Name      *string                      `json:"name,omitempty"`
-	Policyid  *int64                       `json:"policyid,omitempty"`
+	Policyid  *float64                     `json:"policyid,omitempty"`
 	Service   *[]FirewallDoSPolicy6Service `json:"service,omitempty"`
 	Srcaddr   *[]FirewallDoSPolicy6Srcaddr `json:"srcaddr,omitempty"`
 	Status    *string                      `json:"status,omitempty"`
 }
 
 type FirewallDoSPolicy6Anomaly struct {
-	Action           *string `json:"action,omitempty"`
-	Log              *string `json:"log,omitempty"`
-	Name             *string `json:"name,omitempty"`
-	Quarantine       *string `json:"quarantine,omitempty"`
-	QuarantineExpiry *string `json:"quarantine-expiry,omitempty"`
-	QuarantineLog    *string `json:"quarantine-log,omitempty"`
-	Status           *string `json:"status,omitempty"`
-	Threshold        *int64  `json:"threshold,omitempty"`
-	Thresholddefault *int64  `json:"threshold(default),omitempty"`
+	Action           *string  `json:"action,omitempty"`
+	Log              *string  `json:"log,omitempty"`
+	Name             *string  `json:"name,omitempty"`
+	Quarantine       *string  `json:"quarantine,omitempty"`
+	QuarantineExpiry *string  `json:"quarantine-expiry,omitempty"`
+	QuarantineLog    *string  `json:"quarantine-log,omitempty"`
+	Status           *string  `json:"status,omitempty"`
+	Threshold        *float64 `json:"threshold,omitempty"`
+	Thresholddefault *float64 `json:"threshold(default),omitempty"`
 }
 
 type FirewallDoSPolicy6Dstaddr struct {
