+++ v2/models/firewall_identity_based_route.go
@@ -12,7 +12,7 @@ type FirewallIdentityBasedRouteRule struct {
 	Device  *string                                 `json:"device,omitempty"`
 	Gateway *string                                 `json:"gateway,omitempty"`
 	Groups  *[]FirewallIdentityBasedRouteRuleGroups `json:"groups,omitempty"`
-	Id      *int64                                  `json:"id,omitempty"`
+	Id      *float64                                `json:"id,omitempty"`
 }
 
 type FirewallIdentityBasedRouteRuleGroups struct {
